rename ideation service file
