+++ src/app/(main)/my-voyage/[teamId]/ideation/adapters/ideationSA.ts
