+++ src/app/(main)/my-voyage/[teamId]/ideation/components/IdeationForm.tsx
@@ -18,7 +18,7 @@ import {
   type EditIdeationProps,
   addIdeation,
   deleteIdeation,
-} from "@/app/(main)/my-voyage/[teamId]/ideation/adapters/ideationService";
+} from "@/app/(main)/my-voyage/[teamId]/ideation/adapters/ideationSA";
 import useServerAction from "@/hooks/useServerAction";
 import { persistor } from "@/store/store";
 import { onOpenModal } from "@/store/features/modal/modalSlice";
